backend/pyproject.toml
[tool.poetry]
name = "backend"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.13"
alembic = ">=1.16.5,<2.0.0"
fastapi = ">=0.118.0,<0.119.0"
sqlalchemy = ">=2.0.43,<3.0.0"
pydantic-settings = ">=2.11.0,<3.0.0"
uvicorn = ">=0.37.0,<0.38.0"
asyncpg = ">=0.30.0,<0.31.0"
greenlet = ">=3.2.4,<4.0.0"
tenacity = ">=9.1.2,<10.0.0"
pyjwt = ">=2.10.1,<3.0.0"
passlib = ">=1.7.4,<2.0.0"
python-jose = ">=3.5.0,<4.0.0"
python-multipart = ">=0.0.20,<0.0.21"
python-dotenv = ">=1.1.1,<2.0.0"
bcrypt = "==4.3.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
